refactor: move routs to separate fiels
